Escapy

+++ src/com/game/animator/EscapyAnimatorThreadObject.java
@@ -10,12 +10,8 @@ public class EscapyAnimatorThreadObject extends EscapyAnimatorSuper implements R
 	 * Instantiates a new escapy animator thread object.
 	 */
 	public EscapyAnimatorThreadObject() {
-		/** VOID **/
 	}
 
-	/* (non-Javadoc)
-	 * @see java.lang.Runnable#run()
-	 */
 	@Override
 	public void run() {
 		Thread.currentThread().setPriority(Thread.MIN_PRIORITY);
@@ -23,9 +19,8 @@ public class EscapyAnimatorThreadObject extends EscapyAnimatorSuper implements R
 			while (!finished[0]) {
 				for (int i = 0; i < animatedList.size(); i++) {
 					if (animatedList.get(i) instanceof EscapyAnimatorObject) {
-						((EscapyAnimatorObject) animatedList.get(i))
-								.InterruptAnimator((EscapyAnimatorObject) animatedList.get(i));
-						if (indexOfList.get(i).intValue() == EscapyAnimatorSuper.ACTIVE)
+						((EscapyAnimatorObject) animatedList.get(i)).InterruptAnimator((EscapyAnimatorObject) animatedList.get(i));
+						if (indexOfList.get(i) == EscapyAnimatorSuper.ACTIVE)
 							((EscapyAnimatorObject) animatedList.get(i)).defineAnimation();
 					}
 				}
